feat: tour improvements
