+++ packages/raystack/components/tour/tour-context.tsx
@@ -14,24 +14,11 @@ export interface TourContextValue {
   step: TourStep | null;
   open: boolean;
   status: TourStatus;
-  /** Resolved anchor element for the active step; null for detached steps or while waiting. */
   anchor: Element | null;
-  /** Resolved element to spotlight; usually the anchor. */
   spotlightElement: Element | null;
-  /** Whether the popover should be visible (tour open and target resolved). */
   popoverOpen: boolean;
-  /** Tour-level default for hiding the dimmed overlay. */
   disableOverlay: boolean;
-  /** How the spotlight and popover travel between steps. */
   transition: TourTransition;
-  /**
-   * Whether the active step's target is settled and in view. Drives the
-   * fade-in of the spotlight (in every mode) and of the popover (in `fade`
-   * mode) so they appear only once the target has stopped moving, e.g. after
-   * scrolling into view. Computed the same way regardless of transition; in
-   * `move` mode the popover ignores it and glides, but the spotlight still
-   * respects it.
-   */
   revealed: boolean;
   actions: TourActions;
 }
