+++ packages/raystack/components/tour/tour-overlay.tsx
@@ -6,6 +6,7 @@ import { createPortal } from 'react-dom';
 import styles from './tour.module.css';
 import { useTourContext } from './tour-context';
 import { usePrefersReducedMotion } from './use-prefers-reduced-motion';
+import { rectsEqual, type SpotlightRect } from './utils';
 
 export interface TourOverlayProps extends ComponentPropsWithoutRef<'div'> {
   /**
@@ -22,38 +23,6 @@ export interface TourOverlayProps extends ComponentPropsWithoutRef<'div'> {
   spotlightClicks?: boolean;
 }
 
-interface SpotlightRect {
-  x: number;
-  y: number;
-  width: number;
-  height: number;
-}
-
-/**
- * Whether two rects match within half a pixel on every side. Shared with the
- * root's reveal loop, which feeds it `DOMRect`s (`x`/`y` mirror `left`/`top`).
- */
-export function rectsEqual(a: SpotlightRect, b: SpotlightRect) {
-  return (
-    Math.abs(a.x - b.x) < 0.5 &&
-    Math.abs(a.y - b.y) < 0.5 &&
-    Math.abs(a.width - b.width) < 0.5 &&
-    Math.abs(a.height - b.height) < 0.5
-  );
-}
-
-/**
- * Tracks an element's viewport rect with a frame loop, so the spotlight
- * follows through scrolling (any container), resizes and CSS animations — e.g.
- * a dialog's entry transform, which fires no scroll/resize events. Re-renders
- * only when the rect actually changes.
- *
- * When the element changes the previous rect is intentionally kept until the
- * next one is measured, so `move` mode glides between steps instead of blinking
- * and `fade` mode never collapses the hole to `0,0` mid-swap. A disconnected
- * element is skipped (rect frozen at its last good value) rather than collapsed,
- * which would flash the whole screen dim.
- */
 function useSpotlightRect(element: Element | null): SpotlightRect | null {
   const [rect, setRect] = useState<SpotlightRect | null>(null);
 
@@ -81,18 +50,6 @@ function useSpotlightRect(element: Element | null): SpotlightRect | null {
   return element ? rect : null;
 }
 
-/**
- * Cross-fade controller for the spotlight cutout. It holds the currently
- * displayed element (so its hole stays put while it closes) and swaps to the
- * next target only once the tour reports it `revealed`. Because `revealed`
- * already waits out the close (see the root's reveal gate), the reposition lands
- * while the cutout is covered by the dim and is never seen — the hole then fades
- * back open on the exact frame the popover reveals. A vanished target
- * (`target == null`) is dropped at once, so no orphaned hole lingers.
- *
- * `shown` is whether the cutout should currently be open (clear); when it is
- * false the cover patch dims the hole closed while the surrounding dim persists.
- */
 function useSpotlightFade(
   target: Element | null,
   { fade, revealed }: { fade: boolean; revealed: boolean }
@@ -100,9 +57,6 @@ function useSpotlightFade(
   const [displayed, setDisplayed] = useState<Element | null>(target);
 
   useEffect(() => {
-    // Adopt the target when it is settled (revealed), when fade is off, when it
-    // has gone (null), or when nothing is shown yet (first step — it mounts
-    // hidden and then fades in). Otherwise keep the old hole to fade it out.
     if (!fade || target == null || revealed || displayed == null) {
       setDisplayed(target);
     }
@@ -131,14 +85,8 @@ export function TourOverlay({
   } = useTourContext('Tour.Overlay');
   const disabled = step?.disableOverlay ?? disableOverlayTour;
   const reducedMotion = usePrefersReducedMotion();
-  // The spotlight cutout always cross-fades between targets (never slides),
-  // whatever the tour's `transition` mode — that setting only affects the
-  // popover. The dim backdrop itself never disappears once the tour is running;
-  // only the cutout closes and reopens.
   const fade = !reducedMotion;
 
-  // A step is "targeted" when it points at or spotlights an element. Detached
-  // steps (welcome/finish) dim the whole viewport with no cutout.
   const targeted =
     step != null && (step.target != null || step.spotlightTarget != null);
 
@@ -151,10 +99,6 @@ export function TourOverlay({
   const [mounted, setMounted] = useState(false);
   useEffect(() => setMounted(true), []);
 
-  // The dim fades in once, when the first step reveals, and then persists for
-  // the rest of the tour — between steps only the cutout closes and reopens, so
-  // the backdrop never flashes away. Reset on close so it fades in again on the
-  // next open.
   const [entered, setEntered] = useState(false);
   useEffect(() => {
     if (!open || disabled) {
@@ -165,9 +109,7 @@ export function TourOverlay({
   }, [open, disabled, revealed]);
 
   if (!open || !mounted || disabled || !step) return null;
-  // A targeted step with no measured rect is still resolving (or its target
-  // just unmounted). Requiring the *live* target (`spotlightElement`) to exist
-  // is what prevents an orphaned dim from lingering after a target disappears.
+  // Requiring the live `spotlightElement` prevents an orphaned dim after a target unmounts.
   if (targeted && (!spotlightElement || !displayed || !rect)) return null;
 
   const padding = step.spotlightPadding ?? spotlightPadding;
@@ -184,19 +126,17 @@ export function TourOverlay({
         }
       : null;
 
-  // Whether the cutout should currently be open (clear). Between steps it
-  // closes — a cover patch dims the hole shut while the backdrop stays put — and
-  // reopens once the next target settles. Detached steps have no cutout.
   const holeOpen = fade ? shown : true;
 
   return createPortal(
+    // `rest` spread first so consumers can't clobber the chrome attributes below.
     <div
+      {...rest}
       aria-hidden
       data-status={status}
       data-entered={entered ? 'true' : 'false'}
       data-hole-open={holeOpen ? 'true' : 'false'}
       className={cx(styles.overlay, className)}
-      {...rest}
     >
       <div
         className={styles.spotlight}
@@ -209,13 +149,10 @@ export function TourOverlay({
                 height: hole.height,
                 borderRadius: radius
               }
-            : // Detached step: collapse the hole so the shadow dims everything.
-              { left: '50%', top: '50%', width: 0, height: 0 }
+            : { left: '50%', top: '50%', width: 0, height: 0 }
         }
       />
       {hole && (
-        // Fills the cutout with the same dim so the hole can fade shut and back
-        // open without the surrounding backdrop ever disappearing.
         <div
           className={styles.spotlightCover}
           style={{
@@ -229,8 +166,8 @@ export function TourOverlay({
       )}
       {hole ? (
         <>
-          {/* Hit strips block clicks around the hole. The strip over the hole
-              is dropped when spotlightClicks lets clicks reach the target. */}
+          {/* Hit strips block clicks around the hole; the center strip is
+              dropped when spotlightClicks lets clicks through. */}
           <div
             className={styles.overlayHit}
             style={{ top: 0, left: 0, right: 0, height: Math.max(hole.y, 0) }}
