+++ packages/raystack/components/tour/use-prefers-reduced-motion.ts
@@ -2,12 +2,6 @@
 
 import { useEffect, useState } from 'react';
 
-/**
- * Tracks the user's `prefers-reduced-motion` setting, updating if it changes.
- * Defaults to motion-allowed (`false`) on the server and where `matchMedia`
- * is unavailable, so the tour's fade/glide logic degrades to instant swaps for
- * reduced-motion users without any layout-affecting difference on the server.
- */
 export function usePrefersReducedMotion(): boolean {
   const [reduced, setReduced] = useState(false);
